# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Oli Docs — Document Review API",
    "version": "1.0.0",
    "description": "Backend API for the Markdown document review system. All endpoints are under /v1 and (except the Google OAuth callback) require a Cognito JWT bearer token. Roles: admin, revisor."
  },
  "servers": [{ "url": "/v1", "description": "Versioned base path" }],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/reviewers/{id}/deactivate": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "post": {
        "tags": ["Reviewers"],
        "summary": "Deactivate a reviewer (admin)",
        "responses": {
          "200": { "description": "Deactivated", "content": { "application/json": { "schema": { "type": "object", "properties": { "id": { "type": "string" }, "active": { "type": "boolean" } } } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/reviewers/{id}/assignments": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "post": {
        "tags": ["Assignments"],
        "summary": "Assign a module to a reviewer (admin)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateAssignment" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Assignment" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
...
    "/documents/{id}/confirm-upload": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "post": {
        "tags": ["Documents"],
        "summary": "Confirm content landed in S3; starts classification (admin, idempotent)",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Document" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/documents/{id}/module": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "patch": {
        "tags": ["Documents"],
        "summary": "Manually reassign the document's module (admin)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReassignModule" } } } },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Document" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/documents/{id}/approval-policy": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "patch": {
        "tags": ["Documents"],
        "summary": "Set the approval policy (admin)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SetApprovalPolicy" } } } },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Document" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/documents/{id}/force-approval": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "post": {
        "tags": ["Documents"],
        "summary": "Force approval evaluation excluding unsubmitted records (admin)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ForceApproval" } } } },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApprovalResult" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "IdPath": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "ValidationError": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Forbidden": { "description": "Insufficient permissions", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "Conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "enum": ["VALIDATION_ERROR", "UNAUTHENTICATED", "FORBIDDEN", "NOT_FOUND", "CONFLICT", "UPSTREAM_ERROR", "INTERNAL_ERROR"] },
              "message": { "type": "string" },
              "details": { "type": "array", "items": { "type": "object", "properties": { "field": { "type": "string" }, "issue": { "type": "string" } } } }
            }
          }
        }
      },
      "Permission": { "type": "string", "enum": ["read", "edit"] },
...
      "ApprovalPolicy": { "type": "string", "enum": ["all", "any"] },
      "DocumentStatus": { "type": "string", "enum": ["pending_content", "classifying", "classified", "needs_manual_classification", "classification_failed", "in_review", "approved", "rejected"] },
...
      "CreateAssignment": { "type": "object", "required": ["moduleId", "permission"], "properties": { "moduleId": { "type": "string" }, "permission": { "$ref": "#/components/schemas/Permission" } } },
...
      "Assignment": {
        "type": "object",
        "required": ["reviewerId", "moduleId", "permission"],
        "properties": { "reviewerId": { "type": "string" }, "moduleId": { "type": "string" }, "permission": { "$ref": "#/components/schemas/Permission" }, "createdAt": { "type": "string", "format": "date-time" } }
      },
...
      "ReassignModule": { "type": "object", "required": ["moduleId"], "properties": { "moduleId": { "type": "string" } } },
      "SetApprovalPolicy": { "type": "object", "required": ["approvalPolicy"], "properties": { "approvalPolicy": { "$ref": "#/components/schemas/ApprovalPolicy" } } },
      "ForceApproval": {
        "type": "object",
        "required": ["excludedRecordIds", "reason"],
        "properties": { "excludedRecordIds": { "type": "array", "minItems": 1, "items": { "type": "string" } }, "reason": { "type": "string", "maxLength": 500 } }
      },
      "ApprovalResult": {
        "type": "object",
        "properties": { "decision": { "type": "string", "enum": ["approved", "rejected", "pending"] }, "document": { "$ref": "#/components/schemas/Document" } }
      },
      "Document": {
        "type": "object",
        "required": ["id", "name", "s3Key", "status", "approvalPolicy", "createdAt"],
        "properties": {
          "id": { "type": "string" }, "name": { "type": "string" }, "s3Key": { "type": "string" },
          "status": { "$ref": "#/components/schemas/DocumentStatus" },
          "moduleId": { "type": "string" },
          "classification": { "type": "object", "properties": { "source": { "type": "string", "enum": ["ai", "manual"] }, "confidence": { "type": "number" } } },
          "approvalPolicy": { "$ref": "#/components/schemas/ApprovalPolicy" },
          "approvalStatus": { "type": "string", "enum": ["approved", "rejected"] },
          "approvalDecidedAt": { "type": "string", "format": "date-time" },
          "excludedRecords": { "type": "array", "items": { "type": "object", "properties": { "reviewerId": { "type": "string" }, "reason": { "type": "string" } } } },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}